[#21] Move to dir __tests__
