+++ app/src/store/__tests__/buildParams-json-empty.test.js
@@ -2,8 +2,8 @@ import { renderHook } from '@testing-library/react-native';
 /**
  * Mock build json is empty and should use the default state
  */
-jest.mock('../build.json', () => ({}));
-import BuildParamsState from './buildParams';
+jest.mock('../../build.json', () => ({}));
+import BuildParamsState from '../buildParams';
 
 describe('BuildParamsState build.json empty', () => {
   it('should initialize with the correct default state', () => {
