+++ app/src/store/__tests__/buildParams.test.js
@@ -1,5 +1,5 @@
 import { renderHook, act } from '@testing-library/react-native';
-import BuildParamsState from './buildParams';
+import BuildParamsState from '../buildParams';
 
 describe('BuildParamsState', () => {
   it('should initialize with the correct default state from build.json', () => {
